design_introspection: Integrate get_count command with design introspection plugin
