+++ README.md
@@ -7,13 +7,12 @@ This repository contains plugins for
 ## List of plugins
 1. [Design introspection](#design-introspection-plugin)
 2. [FASM](#fasm-plugin)
-3. [Get count](#get-count-plugin)
-4. [Integrate inverters](#integrate-inverters-plugin)
-5. [Parameters](#parameters-plugin) 
-6. [QL IOBs](#quicklogic-iob-plugin)
-7. [SDC](#sdc-plugin)
-8. [Selection](#selection-plugin)
-9. [XDC](#xdc-plugin)
+3. [Integrate inverters](#integrate-inverters-plugin)
+4. [Parameters](#parameters-plugin)
+5. [QL IOBs](#quicklogic-iob-plugin)
+6. [SDC](#sdc-plugin)
+7. [Selection](#selection-plugin)
+8. [XDC](#xdc-plugin)
 
 ## Summary
 
@@ -26,6 +25,7 @@ Following commands are added with the plugin:
 * get_nets
 * get_pins
 * get_ports
+* get_count
 
 ### FASM plugin
 
@@ -33,14 +33,6 @@ Writes out the design's [fasm features](https://symbiflow.readthedocs.io/en/late
 
 The plugin adds the following command:
 * write_fasm
-
-### Get count plugin
-
-Returns the count of selected objects to the TCL interpreter.
-The objects can be of various types, such as modules, cells or wires.
-
-The plugin adds the following command:
-* get_count
 
 ### Integrate inverters plugin
