+++ design_introspection-plugin/Makefile
@@ -4,6 +4,7 @@ SOURCES = design_introspection.cc \
 	  get_nets.cc \
 	  get_ports.cc \
 	  get_cells.cc \
-	  get_pins.cc
+	  get_pins.cc \
+	  get_count.cc
 
 include ../Makefile_plugin.common
