+++ design_introspection-plugin/get_count.h
@@ -17,24 +17,21 @@
  *  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *
  */
+#ifndef _GET_COUNT_H_
+#define _GET_COUNT_H_
 
-#include "get_cells.h"
-#include "get_count.h"
-#include "get_nets.h"
-#include "get_pins.h"
-#include "get_ports.h"
+#include "kernel/register.h"
 
 USING_YOSYS_NAMESPACE
 
-PRIVATE_NAMESPACE_BEGIN
+struct GetCount : public Pass {
 
-struct DesignIntrospection {
-    DesignIntrospection() {}
-    GetNets get_nets_cmd;
-    GetPorts get_ports_cmd;
-    GetCells get_cells_cmd;
-    GetPins get_pins_cmd;
-    GetCount get_count_cmd;
-} DesignIntrospection;
+    enum class ObjectType { NONE, MODULE, CELL, WIRE };
 
-PRIVATE_NAMESPACE_END
+    GetCount() : Pass("get_count", "Returns count of various selected object types to the TCL interpreter") {}
+
+    void help() override;
+    void execute(std::vector<std::string> a_Args, RTLIL::Design *a_Design) override;
+};
+
+#endif // GET_COUNT_H_
