+++ design_introspection-plugin/tests/Makefile
@@ -1,7 +1,8 @@
 TESTS = get_nets \
 	get_ports \
 	get_cells \
-	get_pins
+	get_pins \
+	get_count
 
 UNIT_TESTS = trim_name
 
@@ -11,3 +12,4 @@ get_nets_verify = $(call diff_test,get_nets,txt)
 get_ports_verify = $(call diff_test,get_ports,txt)
 get_cells_verify = $(call diff_test,get_cells,txt)
 get_pins_verify = $(call diff_test,get_pins,txt)
+get_count_verify = true
