--- get_count-plugin/Makefile
@@ -1,3 +1,0 @@
-NAME = get_count
-SOURCES = get_count.cc
-include ../Makefile_plugin.common
