--- get_count-plugin/tests/Makefile
@@ -1,5 +1,0 @@
-TESTS = simple
-
-include $(shell pwd)/../../Makefile_test.common
-
-simple_verify = true
